Add util functions to parser_utils module
