+++ parser/parser.py
@@ -2,42 +2,10 @@
 import sys
 import time
 import os
-import shutil
 import subprocess
 from bs4 import BeautifulSoup
-import xml.dom.minidom
 
-def delete_directory(path):
-    if os.path.exists(path):
-        os.system('rm -rf {}'.format(path))
-        print('Target directory deleted')
-    else:
-        print('Target directory does not exist')
-
-def create_directory(name):
-    try:
-        os.mkdir(name)
-        print(f'{name} directory created')
-    except FileExistsError:
-        print(f'The directory {name} already exists')
-
-def copy_directory(src_folder, dst_folder):
-    shutil.copytree(src_folder, dst_folder)
-
-def copy_files(source_dir, target_dir, file_names):
-    #create the target dir if it doesn't exist
-    if not os.path.exists(target_dir):
-        os.makedirs(target_dir)
-
-    #copy files
-    for file_name in file_names:
-        source_file = os.path.join(source_dir, file_name)
-        target_file = os.path.join(target_dir, file_name)
-        shutil.copy(source_file, target_dir)
-
-def prettify_xml(xml_code):
-    dom = xml.dom.minidom.parseString(xml_code)
-    return dom.toprettyxml()
+from parser.parser_utils import create_directory, delete_directory, copy_directory, copy_files, prettify_xml
 
 def process_regions():
     #copy the world-map.gif file
